requirements.txt
streamlit==1.31.1
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
